+++ compiler-core/src/type_/error.rs
@@ -502,6 +502,7 @@ pub enum UnifyErrorSituation {
     ReturnAnnotationMismatch,
     PipeTypeMismatch,
     Operator(BinOp),
+    InconsistentTry(bool),
 }
 
 impl UnifyErrorSituation {
@@ -519,6 +520,17 @@ annotation of this function.",
                 Some("This function cannot handle the argument sent through the (|>) pipe:")
             }
             Self::Operator(_op) => None,
+            Self::InconsistentTry(return_value_is_at_least_result) => {
+                if *return_value_is_at_least_result {
+                    Some("This returned value has a type incompatible with the previous try expression.
+All the try expressions in a block and the final result value must have
+the same error type.")
+                } else {
+                    Some("This returned value has a type incompatible with the previous try expression.
+The returned value after a try must be of type Result."
+                    )
+                }
+            }
         }
     }
 }
@@ -570,6 +582,12 @@ impl UnifyError {
 
     pub fn operator_situation(self, binop: BinOp) -> Self {
         self.with_unify_error_situation(UnifyErrorSituation::Operator(binop))
+    }
+
+    pub fn inconsistent_try(self, return_value_is_at_least_result: bool) -> Self {
+        self.with_unify_error_situation(UnifyErrorSituation::InconsistentTry(
+            return_value_is_at_least_result,
+        ))
     }
 
     pub fn into_error(self, location: SrcSpan) -> Error {